+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 
 ### :bug: Bug Fix
 
+- `video`
+  - [#3121](https://github.com/wix/ricos/pull/3121) Fix new video modal style
 - `vertical-embed`
   - [#3123](https://github.com/wix/ricos/pull/3123) bug fix scroll to selected item